Rework stream implementation
Previously stream posts depended on the layout and templating system to
do most of the work of generating the final pages. But this made it
difficult to do things like an RSS feed which depend on the final
rendered content. This iteration keeps the functions for generating the
final HTML for different post types, but each post has to call the
corresponding function explicitly.

  - Combine functions for generating HTML for different post types into
    a single file
  - Update existing posts to use functions to generate content, rather
    than leaving it to a layout
  - Simplify post and index pages to pull in rendered content
  - Add generator to paginate stream
  - Update stream index page accordingly
  - Add an RSS feed for stream posts